amd-versal: add i2c device nodes for versal board
Issue: LINUXEXEC-36548

When DISTRO_FEATURES contains "xilinx-bootloader", the device-tree
recipe will accord to the hardware descriptions of xsa files to
create system.dtb based on the dts and dtsi files which downloaded
from:
    https://github.com/Xilinx/system-device-tree-xlnx.git

This means the dts files in linux-yocto will be bypassed. So we add
the i2c dtsi here to cover this scene to ensure that system.dtb will
contain i2c devices.

The reason why i2c buses are not described in linux-xlnx board dts file
is that:

i2c bus on this board (and others) is designed that multiple masters have
access to it. One way is from Versal and second from System Controller.
And because on i2c there are i2c muxes which are not safe for multiple
master access the i2c bus is described only from System Controller point
of view.
And it is up to user to decide if it is safe to use bus for specific i2c
device and also it depends on tasks running on System Controller.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>