amd-versal: add a recipe versal-fw-demo to generate dtbo
Issue: LINUXEXEC-36548

When customers use SDT Segment Configuration designs, a dtsi
file is generated to describe the FPGA implementation. But this
dtsi file must be applied a series of operations to it. The
bbclass "dfx_user_dts" that is provided by AMD will help to do
this. So this recipe is only to pass dtsi and pdi files to
"dfx_user_dts" bbclass to do modications on it.

This recipe refers to the demo in
    https://github.com/Xilinx/meta-xilinx/blob/master/docs/README.dfx.user.dts.md#how-to-create-a-firmware-recipe-app

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>